mount: Set protocol family properly for "udp" and "tcp"
In nfs_nfs_proto_family(), *family is never set if the legacy
"udp" or "tcp" mount options are specified.  The result is an error
message at umount time, for example:

umount.nfs: DNS resolution failed for
  2001:5c0:1101:2f00:250:8dff:fe95:5c61: ai_family not supported

even if mount was built with IPv6 support.

The man page says that "udp" is a synonym for "proto=udp", and
likewise for "tcp".  Thus, we don't look at config_default_family
here, but always use AF_INET explicitly, to be consistent with the
meaning of proto=.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>