nfs-utils: fix default value for --enable-tirpc
We need $enable_tirpc to be a tristate. 'yes' means that someone
explicitly requested building with tirpc. 'no' means that it was
explicitly disabled. Anything else means that no one specified a value.

Fix it by setting the value to a blank string so that the default is
properly undefined.

Reported-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>